uzsu plugin: outsource webif and fix webinterface problem with showing the whole dictionary when a rule contains a "<"
